Fix -Wformat-signedness warnings
Correct the "%d" conversions that "-Wformat-signedness" flags for
unsigned arguments.  GCC gives an enum with no negative enumerators
an unsigned underlying type.  m_msg_type_t, munge_err_t, munge_mac_t,
and munge_cipher_t all qualify, so "%d" mismatches their signedness.
Cast those arguments to int.  "%d" already read them as int, so the
printed text does not change.

Cast rather than switch to "%u".  The C standard leaves an enum's
underlying type implementation-defined.  GCC picks unsigned int here,
but a compiler that picks a signed type would flag "%u" instead.
The cast is correct either way.  "%u" would also print unmunge's "-1"
status sentinel as 4294967295.

"m->data_len" in libmunge's "encode.c" is a real uint32_t.  Use the
PRIu32 conversion there.  This matches how commits c817d0e5 and
d11fa3bc treated other uint32_t message fields.  "-Wformat" alone
does not flag it, because "%d" and a same-width unsigned int differ
only in signedness.

Also correct the format string in "hkdf_rfc_test.c".  Nothing flagged
it.  libtap is vendored source, and its diag() has no format attribute.
The compiler does not check the format strings passed to it.

+++ src/common/hkdf_rfc_test.c
@@ -83,7 +83,7 @@ hkdf_test (munge_mac_t md,
     }
     rv = hkdf_ctx_set_md (hkdfp, md);
     if (rv == -1) {
-        diag ("Failed to set HKDF message digest to md=%d", md);
+        diag ("Failed to set HKDF message digest to md=%d", (int) md);
         return -1;
     }
     if (ikm != NULL) {
